Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-analysernode-interface/test-analysernode-expected.txt


FAIL Test AnalyserNode API assert_equals: analyser node has 1 input channels by default expected 1 but got 2
FAIL Test AnalyserNode's ctor API function is not a constructor (evaluating 'new AnalyserNode(context)')
